move test and implementation to `cli`
