+++ packages/cli/vitest.config.ts
@@ -1,0 +1,11 @@
+import { defineProject } from 'vitest/config';
+
+export default defineProject({
+	test: {
+		name: 'cli',
+		include: ['./tests/**/index.ts', './tests/*.ts'],
+		expect: {
+			requireAssertions: true
+		}
+	}
+});
